fix testing environment issue
